Change mssql tinyint to smallint to match data types used in postgres. Change PurgeBatchSize from int to short in ScheduleRepositoryConfiguration to sql data type.

+++ DatabaseScripts/EasyNetQ.Scheduler/mssql/CreateWorkTables.sql
@@ -19,7 +19,7 @@ GO
 
 CREATE TABLE [dbo].[WorkItemStatus] (
 	[WorkItemID]  [int] NOT NULL,
-	[Status] [tinyint] NULL,
+	[Status] [smallint] NULL,
 	[WakeTime] [datetime] NULL,
 	[ClientID] [tinyint] NULL,
 	[PurgeDate] [smalldatetime] NULL
